Change average han by player chart
Fixes #37

Replace the `AverageHanChart` component with a new `AverageHanTable` component to display average han values in a table format.

* **Add `AverageHanTable` component:**
  - Create `AverageHanTable` component in `src/components/PlayerScoreChart/AverageHanTable.tsx`
  - Use MUI components (`Table`, `TableHead`, `TableBody`, `TableRow`, `TableCell`) to display average han values for each player
  - Style table cells with player colors

* **Update `PlayerScoreChart` component:**
  - Import `AverageHanTable` from `./PlayerScoreChart/AverageHanTable`
  - Replace `AverageHanChart` component with `AverageHanTable` component

* **Remove `AverageHanChart` component:**
  - Delete `src/components/PlayerScoreChart/AverageHanChart.tsx`

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/josve/mahjong/issues/37?shareId=XXXX-XXXX-XXXX-XXXX).
